# employee-management/docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controllers.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  models.Employee:
    properties:
      address:
        type: string
      cmnd:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      department_id:
        type: integer
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      position_id:
        type: integer
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Salary:
    properties:
      basic_salary:
        type: integer
      bonus:
        type: integer
      coefficient:
        type: integer
      created_at:
        type: string
      employee:
        allOf:
        - $ref: '#/definitions/models.Employee'
        description: Liên kết với nhân viên (Employee)
      employee_id:
        description: Khóa ngoại liên kết với bảng Employee
        type: integer
      employee_name:
        type: string
      fine:
        type: integer
      id:
        type: integer
      total_salary:
        type: integer
      updated_at:
        type: string
      working_days:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing employees, departments, positions, and salaries.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Employee Management API
  version: "1.0"
paths:
...
  /api/v1/salaries:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all salaries, including employee information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Salary'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get all salaries
      tags:
      - Salary
    post:
      consumes:
      - application/json
      description: Add a new salary record for an employee
      parameters:
      - description: Salary data
        in: body
        name: salary
        required: true
        schema:
          $ref: '#/definitions/models.Salary'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Salary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new salary
      tags:
      - Salary
  /api/v1/salaries/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a salary record by its ID
      parameters:
      - description: Salary ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete a salary
      tags:
      - Salary
    get:
      consumes:
      - application/json
      description: Retrieve a salary by its ID, including employee information
      parameters:
      - description: Salary ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Salary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get salary by ID
      tags:
      - Salary
    put:
      consumes:
      - application/json
      description: Update a salary record by its ID
      parameters:
      - description: Salary ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated salary data
        in: body
        name: salary
        required: true
        schema:
          $ref: '#/definitions/models.Salary'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Salary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Update an existing salary
      tags:
      - Salary
...
swagger: "2.0"